modificando controladores para poder subir varias imagenes y distinguirlas de la imagen principal para renderizarlas en el productDetail luego

+++ controllers/productsController.js
@@ -18,17 +18,24 @@ const productsController = {
     }, 
     store: (req, res)=>{
 
+        //const mainImage = req.files.file_img[0];
+       
         const nuevoId = products ? products[products.length - 1].id + 1 : 0 ;
-        const image = req.file ? req.file.filename : 'image.png';
+        //const image = req.files.file_img > 0 ? mainImage.filename : 'image.png';
+        const image = req.files.file_img[0].filename ? req.files.file_img[0].filename : 'image.png';
+        const secondaryImages = req.files.files_img;
+        //console.log(secondaryImages)
+        console.log(req.body)
 
         const newProduct = {
             id: nuevoId , 
             ...req.body , 
-            imagen: image
+            imagen: image,  
+            imagenes: secondaryImages ? secondaryImages : [{filename: 'image.png' }]
         }
 
         products.push(newProduct);
-        fs.writeFileSync( path.join(__dirname , '../data/products.JSON') , JSON.stringify(products) )
+        fs.writeFileSync( path.join(__dirname , '../data/products.JSON') , JSON.stringify(products, null, 2) )
         res.redirect('/products')
     }, 
 
@@ -45,16 +52,21 @@ const productsController = {
                 if (req.file) {
                    fs.unlinkSync(path.join(__dirname , `../public/products/${products[i].imagen}`))
                 } 
-                const image = req.file ? req.file.filename : 'image.png';
+                //const image = req.files.file_img > 0 ? req.files.file_img[0].filename : 'image.png';
+                const image = req.files.file_img[0].filename ? req.files.file_img[0].filename : 'image.png';
+
+                const secondaryImages = req.files.files_img;
+                console.log(req.body)
                 products[i] = {
                     id: req.params.id, 
                     ...req.body, 
-                    imagen: image
+                    imagen: image, 
+                    imagenes: secondaryImages ? secondaryImages : [{filename: 'image.png' }]
+
                 }; 
             }
         };
-
-        fs.writeFileSync(  path.join(__dirname , '../data/products.JSON') , JSON.stringify(products)); 
+        fs.writeFileSync(  path.join(__dirname , '../data/products.JSON') , JSON.stringify(products, null, 2)); 
         res.redirect('/products'); 
     }, 
 
@@ -62,9 +74,19 @@ const productsController = {
 
         const productsFiltrados = products.filter( p => p.id != req.params.id ); 
         const elegido = products.find( p => p.id == req.params.id );
-        fs.unlinkSync(path.join(__dirname , `../public/products/${elegido.imagen}`));
+
+        if (elegido.imagen =! 'image.png' ) {
+            fs.unlinkSync(path.join(__dirname , `../public/products/${elegido.imagen}`));
+        }
+
+        if ( elegido.imagenes.length > 0 ) {
+            for (let i = 0; i < elegido.imagenes.length; i++) {
+                fs.unlinkSync(path.join(__dirname , `../public/products/${ elegido.imagenes[i].filename }`));
+            }
+        }
+
         products = [...productsFiltrados]; 
-        fs.writeFileSync(  path.join(__dirname , '../data/products.JSON') , JSON.stringify(products)); 
+        fs.writeFileSync(  path.join(__dirname , '../data/products.JSON') , JSON.stringify(products , null, 2)); 
         res.redirect('/products'); 
     } 
 }
